Update and rename useKeyboardShortcuts.js to useKeyboardShortcuts.ts
